+++ tests/unit/models/gpflow/test_config.py
@@ -21,15 +21,15 @@ import gpflow
 import numpy as np
 import pytest
 import tensorflow as tf
-from gpflow.models import GPR, SGPR, VGP, SVGP, GPMC
+from gpflow.models import GPMC, GPR, SGPR, SVGP, VGP
 
-from tests.util.models.gpflow.models import fnc_3x_plus_10, gpr_model, svgp_model
+from tests.util.models.gpflow.models import fnc_3x_plus_10, gpr_model
 from trieste.models import TrainableProbabilisticModel
 from trieste.models.gpflow import (
     GaussianProcessRegression,
     GPflowModelConfig,
-    VariationalGaussianProcess,
     SparseVariational,
+    VariationalGaussianProcess,
 )
 from trieste.models.optimizer import Optimizer
 
@@ -37,9 +37,7 @@ from trieste.models.optimizer import Optimizer
 def test_gpflow_model_config_raises_not_supported_model_type() -> None:
     x = tf.constant(np.arange(5).reshape(-1, 1), dtype=gpflow.default_float())
     y = fnc_3x_plus_10(x)
-    model_specs = {
-        "model": GPMC((x, y), gpflow.kernels.Matern32(), gpflow.likelihoods.Gaussian())
-    }
+    model_specs = {"model": GPMC((x, y), gpflow.kernels.Matern32(), gpflow.likelihoods.Gaussian())}
 
     with pytest.raises(NotImplementedError):
         GPflowModelConfig(**model_specs)
