+++ tests/unit/models/gpflow/test_models.py
@@ -27,14 +27,15 @@ from __future__ import annotations
 
 import unittest.mock
 from collections.abc import Iterable
+from typing import Any
 
 import gpflow
-from gpflow.models import GPR, SGPR, VGP, SVGP
 import numpy as np
 import numpy.testing as npt
 import pytest
 import tensorflow as tf
 import tensorflow_probability as tfp
+from gpflow.models import SGPR, SVGP, VGP
 
 from tests.util.misc import random_seed
 from tests.util.models.gpflow.models import (
@@ -42,8 +43,8 @@ from tests.util.models.gpflow.models import (
     fnc_2sin_x_over_3,
     fnc_3x_plus_10,
     gpr_model,
-    sgpr_model,
     mock_data,
+    sgpr_model,
     svgp_model,
     vgp_matern_model,
     vgp_model,
@@ -54,7 +55,8 @@ from trieste.models.gpflow import (
     SparseVariational,
     VariationalGaussianProcess,
 )
-from trieste.models.optimizer import Optimizer, TFOptimizer, create_optimizer
+from trieste.models.optimizer import DatasetTransformer, Optimizer, TFOptimizer, create_optimizer
+from trieste.types import TensorType
 
 
 def _3x_plus_gaussian_noise(x: tf.Tensor) -> tf.Tensor:
@@ -185,7 +187,10 @@ def test_gpr_raises_for_invalid_num_kernel_samples() -> None:
 )
 @pytest.mark.parametrize("prior_for_lengthscale", [True, False])
 def test_gaussian_process_regression_correctly_counts_params_that_can_be_sampled(
-    mocked_model_initializer, dim, prior_for_lengthscale, gpr_interface_factory: ModelFactoryType
+    mocked_model_initializer: Any,
+    dim: int,
+    prior_for_lengthscale: bool,
+    gpr_interface_factory: ModelFactoryType,
 ) -> None:
     x = tf.constant(np.arange(1, 5 * dim + 1).reshape(-1, dim), dtype=tf.float64)  # shape: [5, d]
     model, _ = gpr_interface_factory(x, fnc_3x_plus_10(x))
@@ -481,7 +486,6 @@ def test_variational_gaussian_process_predict() -> None:
     npt.assert_allclose(variance_y - model.get_observation_noise(), variance, atol=5e-5)
 
 
-
 def test_sparse_variational_model_attribute() -> None:
     model = svgp_model(*mock_data())
     sv = SparseVariational(model)
@@ -520,7 +524,7 @@ def test_sparse_variational_optimize_with_defaults() -> None:
     assert model.model.training_loss(data) < loss
 
 
-def _batcher_1(dataset: Dataset, batch_size: int) -> Iterable:
+def _batcher_1(dataset: Dataset, batch_size: int) -> Iterable[tuple[TensorType, TensorType]]:
     ds = tf.data.Dataset.from_tensor_slices(dataset.astuple())
     ds = ds.shuffle(100)
     ds = ds.batch(batch_size)
@@ -528,12 +532,12 @@ def _batcher_1(dataset: Dataset, batch_size: int) -> Iterable:
     return iter(ds)
 
 
-def _batcher_2(dataset: Dataset, batch_size: int):
+def _batcher_2(dataset: Dataset, batch_size: int) -> tuple[TensorType, TensorType]:
     return dataset.astuple()
 
 
 @pytest.mark.parametrize("batcher", [_batcher_1, _batcher_2])
-def test_sparse_variational_optimize(batcher, compile: bool) -> None:
+def test_sparse_variational_optimize(batcher: DatasetTransformer, compile: bool) -> None:
     x_observed = np.linspace(0, 100, 100).reshape((-1, 1))
     y_observed = _3x_plus_gaussian_noise(x_observed)
     data = x_observed, y_observed
@@ -551,7 +555,9 @@ def test_sparse_variational_optimize(batcher, compile: bool) -> None:
 
 @pytest.mark.parametrize("use_natgrads", [True, False])
 @pytest.mark.parametrize("batcher", [_batcher_1, _batcher_2])
-def test_vgp_optimize_with_and_without_natgrads(batcher, compile: bool, use_natgrads: bool) -> None:
+def test_vgp_optimize_with_and_without_natgrads(
+    batcher: DatasetTransformer, compile: bool, use_natgrads: bool
+) -> None:
     x_observed = np.linspace(0, 100, 100).reshape((-1, 1))
     y_observed = _3x_plus_gaussian_noise(x_observed)
     data = x_observed, y_observed
